% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Annotation.R
\name{Cell_Annotation}
\alias{Cell_Annotation}
\title{Title}
\usage{
Cell_Annotation(
  x,
  method = "SingleR",
  ref,
  label,
  cluster = TRUE,
  marker = NULL,
  weight_id = "avg_log2FC",
  scale = TRUE
)
}
\arguments{
\item{x}{a seurat object}

\item{method}{a character, only can set to "SingleR" in this version}

\item{ref}{reference single cell RNAseq data}

\item{label}{reference labels}

\item{cluster}{logical value, whether to label cells by cluster}
}
\value{
a seurat object with predicted cell label
}
\description{
Title
}
