% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Comunication.R
\name{Cell_Communication}
\alias{Cell_Communication}
\title{Cell_Communication}
\usage{
Cell_Communication(pathlist, rm_complex = FALSE, top = 20)
}
\arguments{
\item{pathlist}{a list containing the paths to CellphoneDB outdir}

\item{rm_complex}{logical value, infer whether delete the complex interaction in the final communication}

\item{top}{number, the top n ligand-receptor interaction in each cluster to be presented.}
}
\value{
a matrix, containing the ligand-receptor enrichment value
}
\description{
Cell_Communication
}
