% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CellRatio_bar}
\alias{CellRatio_bar}
\title{CellRatio_bar}
\usage{
CellRatio_bar(x, meta = NULL)
}
\arguments{
\item{x}{spatalk object from Standard_Spatalk or a seurat object
(this function will calculate the idents of the seurat object's ratio)}
}
\description{
CellRatio_bar
}
