% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Extract_Target}
\alias{Extract_Target}
\title{Extract_Target}
\usage{
Extract_Target(x, celltype_sender, celltype_receiver, top = 20)
}
\arguments{
\item{x}{a SpaTalk object}

\item{celltype_sender}{name of sender cell}

\item{celltype_receiver}{name of receiver cell}

\item{top}{top n ligand-receptor pairs to be extract}
}
\value{
a series of characters containing targeted genes in the given sender and receiver cells
}
\description{
Extract_Target
}
