% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Fun_Plot}
\alias{Fun_Plot}
\title{Fun_Plot}
\usage{
Fun_Plot(
  x,
  species,
  showCategory = 20,
  celltype_sender = NULL,
  celltype_receiver = NULL,
  top = 20
)
}
\arguments{
\item{x}{a SpaTalk object or data.frame with gene and cluster information or a series of genes}

\item{species}{name of species}

\item{showCategory}{number of GO/KEGG categories to be shown}

\item{celltype_sender}{name of sender cell}

\item{celltype_receiver}{name of receiver cell}

\item{top}{top n ligand-receptor pairs to be extract}
}
\value{
a list containing GO and KEGG output
}
\description{
Fun_Plot
}
