% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{LRpair_enrich}
\alias{LRpair_enrich}
\title{LRpair_enrich}
\usage{
LRpair_enrich(
  object,
  celltype_sender,
  celltype_receiver,
  top_lrpairs = 20,
  color = NULL,
  border_color = "black",
  type = "number",
  fontsize_number = 5,
  number_color = "black",
  color_low = NULL,
  color_high = NULL
)
}
\arguments{
\item{object}{SpaTalk object}

\item{celltype_sender}{name of sender cell}

\item{celltype_receiver}{name of receiver cell}

\item{top_lrpairs}{number of top lrpairs for plotting}

\item{color}{color for the cells in heatmap.}

\item{border_color}{color of cell borders on heatmap, use NA if no border should be drawn.}

\item{type}{recommend set to "number", plot significant LR with number as fill color}

\item{fontsize_number}{fontsize of the numbers displayed in cells}

\item{number_color}{color of the text}

\item{color_low}{for 'number' type, define the color for the lowest value.}

\item{color_high}{for 'number' type, define the color for the highest value}
}
\description{
LRpair_enrich
}
