% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mark_Nuclei.R
\name{Mark_Nuclei}
\alias{Mark_Nuclei}
\title{Mark_Nuclei}
\usage{
Mark_Nuclei(path, QC_dir = NULL, project)
}
\arguments{
\item{path}{a character infer the parent directory to cellranger filtered_feature_bc_matrix}

\item{QC_dir}{a character indicate the path you want save your QC figures. Make sure you have access to the QC directory}

\item{project}{a character indicate your sample name for identification}
}
\value{
a seurat object within nuclei information in meta.data
}
\description{
Mark_Nuclei
}
