% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpArea_Select.R
\name{SpArea_Select}
\alias{SpArea_Select}
\title{SpArea_Select
this function is to select boundary spot manually.}
\usage{
SpArea_Select(
  x,
  ident,
  group,
  interestsID,
  spatial10x = FALSE,
  saveCPDB_dir = NULL
)
}
\arguments{
\item{x}{a seurat object}

\item{ident}{ident to label the subset object (eg. "xx_boundary")}

\item{group}{group to cluster your data, should be in your meta.data (eg. celltype slot in your seurat meta.data)}

\item{interestsID}{characters infer your interested spot to select, the character should be in the group above.}

\item{spatial10x}{logical value to infer if the data is from cellranger output}

\item{saveCPDB_dir}{directory path to save the files for CellphoneDB}
}
\value{
a seurat object within selected spots, present the selected area in the plot
}
\description{
SpArea_Select
this function is to select boundary spot manually.
}
