% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard_Seu.R
\name{Standard_Seu}
\alias{Standard_Seu}
\title{Standard_Seu}
\usage{
Standard_Seu(x, spatial_dir = NULL, QC_dir = NULL, project, filter = TRUE)
}
\arguments{
\item{x}{a seurat object, a count matrix or a parent directory path to cellranger filtered_feature_bc_matrix}

\item{spatial_dir}{the cellranger out spatial directory. Due to lack of spatial HE images, if your data is from stereoseq, this parameter should set to NULL.}

\item{QC_dir}{if not null, should be the QC outdir path you want save your QC figures. Make sure you have access to the QC directory}

\item{project}{your sample name for identification}

\item{filter}{logical, whether to filter cells/spots in top5\% in lowest minfeature and highest maxfeature with top5\% mitochondrial genes.}
}
\value{
a seurat object
}
\description{
Standard_Seu
}
