% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard_Spatalk.R
\name{Standard_Spatalk}
\alias{Standard_Spatalk}
\title{Standard_Spatalk}
\usage{
Standard_Spatalk(
  sp,
  sc,
  sc_celltype,
  method = "SpaTalk",
  sp_meta = NULL,
  geneinfo = NULL,
  communication = FALSE,
  lrpairs = NULL,
  pathways = NULL
)
}
\arguments{
\item{sp}{a seurat object or count matrix  of spatial transcriptome data}

\item{sc}{a seurat object or count matrix of reference scRNAseq data}

\item{sc_celltype}{a series character label of the sc reference's cell barcodes}

\item{sp_meta}{a matrix recording the spatial locate information, column as x, y and rownames are spot barcodes}

\item{geneinfo}{the same as description in SpaTalk, Null, to use default geneinfo in spatalk}

\item{lrpairs}{the same as description in SpaTalk}

\item{pathways}{the same as description in SpaTalk}
}
\value{
a spatalk object with celltype and ligand-receptor enrichment information
}
\description{
Standard_Spatalk
}
